Add files via upload
I have made this model which will try to predict estimated price of old car on basis of thier features. As now a day we know many people are going to buy second hand car instead of buying new one, but main question is how will us know actual price of car on their features so in order to solve this problem I have used this dataset to build model which will give a estimated price of car at which car should be sold.